refactor(arcade): isolate blackjack controls
